requirements.txt
espn-api
pybaseball
requests
anthropic
playwright
python-dotenv
